+++ packages/elements/src/components/ui/label/Label.tsx
@@ -1,12 +1,14 @@
 import { Box, BoxProps, Indent, Row, Space, Text } from "@stenajs-webui/core";
 import { cssColor } from "@stenajs-webui/theme";
 import * as React from "react";
+import { ComponentPropsWithoutRef } from "react";
 
 export interface LabelProps extends BoxProps {
   text: string;
   optional?: boolean;
   maxCharacters?: number;
   textWidth?: BoxProps["width"];
+  htmlFor?: ComponentPropsWithoutRef<"label">["htmlFor"];
 }
 
 export const Label: React.FC<LabelProps> = React.memo(
@@ -17,7 +19,8 @@ export const Label: React.FC<LabelProps> = React.memo(
     maxCharacters,
     row,
     textWidth,
-    ...columnProps
+    htmlFor,
+    ...boxProps
   }) => {
     const extraInfoLabel = getExtraInfoLabel(optional, maxCharacters);
 
@@ -32,8 +35,8 @@ export const Label: React.FC<LabelProps> = React.memo(
     );
 
     return (
-      <label>
-        <Box {...columnProps} row={row}>
+      <label htmlFor={htmlFor}>
+        <Box {...boxProps} row={row}>
           <Row alignItems={"center"} width={textWidth}>
             <Text
               color={cssColor("--lhds-color-ui-600")}